requirements.txt
# Honeyclaw Dependencies

# Core SSH Honeypot
asyncssh>=2.14.0

# HTTP client for threat intel enrichment
aiohttp>=3.9.0

# Configuration parsing
pyyaml>=6.0

# Optional: Web framework for API honeypots
aiohttp

# Development
pytest>=7.0
pytest-asyncio>=0.21
